+++ src/soho/accordion/soho-accordion.d.ts
@@ -60,6 +60,12 @@ interface SohoAccordionStatic {
   /** Enables the accordion. */
   enable(): void;
 
+  /** Checks if a particular header is disabled, or if the entire accordion is disabled. */
+  isDisabled(jQuery): boolean;
+
+  /** Checks if an Accordion Section is currently expanded. */
+  isExpanded(jQuery): boolean;
+
   /** Toggles the exanded state of the selected header. */
   toggle(jQuery): void;